Moved stuff to mol-util/json.ts
