+++ src/mol-util/_spec/json.spec.ts
@@ -4,22 +4,10 @@
  * @author Adam Midlik <midlik@gmail.com>
  */
 
-import { canonicalJsonString, sortObjectKeys } from '../object';
+import { canonicalJsonString } from '../json';
 
 
 describe('object utils', () => {
-    it('sortObjectKeys', async () => {
-        expect(sortObjectKeys({})).toEqual({});
-
-        const obj1 = { c: 1, b: 2, d: undefined, a: { f: undefined, e: 4 } };
-        expect(sortObjectKeys(obj1)).toEqual(obj1);
-        expect(Object.keys(sortObjectKeys(obj1))).toEqual(['a', 'b', 'c']);
-
-        const obj2 = { c: [], x: null, b: false, a: undefined };
-        expect(sortObjectKeys(obj2)).toEqual(obj2);
-        expect(Object.keys(sortObjectKeys(obj2))).toEqual(['b', 'c', 'x']);
-    });
-
     it('canonicalJsonString', async () => {
         expect(canonicalJsonString({})).toEqual('{}');
 
@@ -30,4 +18,3 @@ describe('object utils', () => {
         expect(canonicalJsonString(obj2)).toEqual('{"b":false,"c":[1,{"p":"P"},0],"x":null}');
     });
 });
-
