+++ src/mol-util/object.ts
@@ -161,20 +161,8 @@ export function objectFromKeysAndValues<K extends keyof any, V>(keys: K[], value
 }
 
 /** Decide if `obj` is a good old object (not array or null or other type). */
-export function isReallyObject(obj: any): boolean {
+export function isPlainObject(obj: any): boolean {
     return typeof obj === 'object' && obj !== null && !Array.isArray(obj);
-}
-
-/** Return a copy of object `obj` with sorted keys and dropped keys whose value is undefined. */
-export function sortObjectKeys<T extends {}>(obj: T): T {
-    const result = {} as T;
-    for (const key of Object.keys(obj).sort() as (keyof T)[]) {
-        const value = obj[key];
-        if (value !== undefined) {
-            result[key] = value;
-        }
-    }
-    return result;
 }
 
 /** Like `Promise.all` but with objects instead of arrays */
@@ -183,20 +171,4 @@ export async function promiseAllObj<T extends {}>(promisesObj: { [key in keyof T
     const promises = Object.values(promisesObj);
     const results = await Promise.all(promises);
     return objectFromKeysAndValues(keys, results) as any;
-}
-
-
-/** A JSON-serializable value */
-export type Jsonable = string | number | boolean | null | Jsonable[] | { [key: string]: Jsonable | undefined }
-
-/** Return a canonical string representation for a JSON-able object,
- * independent from object key order and undefined properties. */
-export function canonicalJsonString(obj: Jsonable) {
-    return JSON.stringify(obj, (key, value) => isReallyObject(value) ? sortObjectKeys(value) : value);
-}
-
-/** Return a pretty JSON representation for a JSON-able object,
- * (single line, but use space after comma). E.g. '{"name": "Bob", "favorite_numbers": [1, 2, 3]}' */
-export function onelinerJsonString(obj: Jsonable) {
-    return JSON.stringify(obj, undefined, '\t').replace(/,\n\t*/g, ', ').replace(/\n\t*/g, '');
 }
